refactor: make the left panel a registry
The panel held an options array and one hardcoded v-show block per tab. Both
lists had to stay in sync by hand.

Extract BlocksTab, AssetsTab, LayersTab and CodeTab into
components/LeftPanelTabs. Each tab owns the padding the panel template
applied to it.

LayersTab takes over activeLayers, the two BlockLayers refs, the breakpoint
header, both DOM class watchers and the clear-selection click. The shell
cannot be generic while it reads BlockLayers refs.

Variables is not a tab. It registers an action and its own isActive, so the
shell needs no special case.

mountedTabs replaces pageScriptMounted and generalizes it. A lazy tab also
declares preload, which the Code tab uses to mount for a data script dialog
it never opens.

Widen LeftSidebarTabOption with (string & {}) so an extension can add a tab.

Part of extensibility phase 1, step 8.
