+++ frontend/src/builder.d.ts
@@ -103,7 +103,14 @@ declare type HashString = `#${string}`;
 
 declare type RGBString = `rgb(${number}, ${number}, ${number})`;
 
-declare type LeftSidebarTabOption = "Blocks" | "Layers" | "Assets" | "Code" | "variables";
+// the literals keep autocomplete; the string widens it for extension tabs
+declare type LeftSidebarTabOption =
+	| "Blocks"
+	| "Layers"
+	| "Assets"
+	| "Code"
+	| "variables"
+	| (string & {});
 
 declare type BuilderMode = "select" | "text" | "container" | "image" | "repeater" | "move";
