+++ frontend/src/components/LeftPanelTabs/AssetsTab.vue
@@ -1,0 +1,6 @@
+<template>
+	<BuilderAssets class="px-3 pb-3" />
+</template>
+<script setup lang="ts">
+import BuilderAssets from "@/components/BuilderAssets.vue";
+</script>
