+++ frontend/src/components/LeftPanelTabs/BlocksTab.vue
@@ -1,0 +1,6 @@
+<template>
+	<BuilderBlockTemplates class="px-3 pb-3" />
+</template>
+<script setup lang="ts">
+import BuilderBlockTemplates from "@/components/BuilderBlockTemplates.vue";
+</script>
